09-serverless/requirements.txt
numpy
tflite_runtime==2.14
Pillow